+++ utils/utils/esbuild.js
@@ -1,0 +1,14 @@
+const esbuild = require('esbuild');
+
+esbuild
+  .build({
+    entryPoints: ['src/index.ts'],
+    outdir: 'lib',
+    bundle: true,
+    sourcemap: true,
+    minify: true,
+    splitting: true,
+    format: 'esm',
+    target: ['esnext']
+  })
+  .catch(() => process.exit(1));