+++ utils/types/CHANGELOG.md
@@ -1,3 +1,19 @@
+<a name="2.0.1"></a>
+
+# 2.0.1 (2021-05-25)
+
+### Fix
+
+- js output needed for enums
+
+<a name="2.0.0"></a>
+
+# 2.0.0 (2021-05-25)
+
+### Build
+
+- build only type declarations
+
 <a name="1.1.0"></a>
 
 # 1.1.0 (2020-04-10)